Official launch of CodeParty

+++ frontend/src/firebase-client/useDeleteOwnAccount.ts
@@ -2,6 +2,7 @@ import { auth } from "./firebase_config";
 import { AuthContext } from "../contexts/AuthContext";
 import { useContext } from "react";
 import { userApiPathAddress } from "@/gateway-address/gateway-address";
+import { GithubAuthProvider, reauthenticateWithCredential, signInWithPopup } from "firebase/auth";
 
 export const useDeleteOwnAccount = () => {
   const { dispatch } = useContext(AuthContext);
@@ -10,7 +11,14 @@ export const useDeleteOwnAccount = () => {
       const currentUser = auth.currentUser;
       if (currentUser) {
         const idToken = await currentUser.getIdToken(true);
-
+        const provider = new GithubAuthProvider();
+        const oAuthResult = await signInWithPopup(auth, provider);
+        const credential = GithubAuthProvider.credentialFromResult(oAuthResult);
+        if (!credential) {
+          throw new Error("Could not get credential");
+        }
+        await reauthenticateWithCredential(currentUser, credential);
+        
         await fetch(userApiPathAddress + currentUser.uid, {
           method: "DELETE",
           headers: {
